First commit to extrapolation branch with general maintenance for Register renamer & Renamed decoder levels and changed the level map
I've fixed the random instruction generation, modified the logic to require the player to not rename 'zr' & allowed to solve the level without using a RAM component & instead requiring the use of memory probes in the `symphony_register_renamer` level;; corrected the test logic in the `symphony_decoder_2` level; and added/corrected the `Level log`-s to the both of them (both of which need a general sense pass). I also changed the level map to make the fixed levels available while disconnecting the one's that are incomplete.